+++ app/components/businesses/hiring_fee_component.rb
@@ -21,11 +21,11 @@ module Businesses
     end
 
     def subject
-      "I hired a developer via railsdevs"
+      t(".subject")
     end
 
     def body
-      "Hi Joe, I hired #{developer} via railsdevs and would like to coordinate paying the fee. Their starting salary is: "
+      t(".body", developer:)
     end
   end
 end