+++ app/components/nav_bar/guest_component.rb
@@ -1,7 +1,9 @@
-class NavBar::GuestComponent < ApplicationComponent
-  attr_reader :links
+module NavBar
+  class GuestComponent < ApplicationComponent
+    attr_reader :links
 
-  def initialize(links:)
-    @links = links
+    def initialize(links:)
+      @links = links
+    end
   end
 end