+++ app/components/users/paywalled_component.rb
@@ -1,9 +1,11 @@
 module Users
   class PaywalledComponent < ApplicationComponent
-    def initialize(user:, paywalled:, size: nil)
+    def initialize(user:, paywalled:, size: nil, title: nil, description: nil)
       @user = user
       @paywalled = paywalled
       @size = size
+      @title = title
+      @description = description
     end
 
     def render_content?
@@ -16,6 +18,14 @@ module Users
 
     def large?
       @size == :large
+    end
+
+    def title
+      @title || t(".title")
+    end
+
+    def description
+      @description || t(".description")
     end
 
     private