+++ test/components/businesses/open_graph_tags_component_test.rb
@@ -1,6 +1,8 @@
 require "test_helper"
 
 class OpenGraphTagsComponentTest < ViewComponent::TestCase
+  include TurboNativeHelper
+
   test "present attributes" do
     render_inline OpenGraphTagsComponent.new
 
@@ -19,5 +21,14 @@ class OpenGraphTagsComponentTest < ViewComponent::TestCase
 
     assert_meta property: "og:title", content: "Custom title · railsdevs"
     assert_meta property: "og:description", content: "And a custom description."
+  end
+
+  test "Turbo Native attributes" do
+    turbo_native_request!
+    render_inline OpenGraphTagsComponent.new(
+      turbo_native_title: "Turbo Native title"
+    )
+
+    assert_meta property: "og:title", content: "Turbo Native title"
   end
 end