+++ test/integration/read_notifications_test.rb
@@ -3,7 +3,7 @@ require "test_helper"
 class ReadNotificationsTest < ActionDispatch::IntegrationTest
   test "you must be signed in" do
     get read_notifications_path
-    assert_redirected_to new_user_registration_path
+    assert_redirected_to new_user_session_path
   end
 
   test "you can view the history page even if no read notifications exist" do
@@ -15,7 +15,7 @@ class ReadNotificationsTest < ActionDispatch::IntegrationTest
 
   test "you can view your past notifications if you have past (read) notifications" do
     sign_in users(:subscribed_business)
-    notifications(:message).mark_as_read!
+    notifications(:message_to_business).mark_as_read!
 
     get read_notifications_path
 
@@ -24,7 +24,7 @@ class ReadNotificationsTest < ActionDispatch::IntegrationTest
 
   test "you can mark all unread notifications as read" do
     sign_in users(:subscribed_business)
-    notification = notifications(:message)
+    notification = notifications(:message_to_business)
 
     refute notification.read?
     post read_notifications_path