+++ test/support/helpers/turbo_native_helper.rb
@@ -1,0 +1,5 @@
+module TurboNativeHelper
+  def turbo_native_request!
+    request.user_agent = "Turbo Native"
+  end
+end